create scroll componet and organize folder
