+++ src/index.css
@@ -5,16 +5,16 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background: #11998e; /* fallback for old browsers */
+  background: #9796f0; /* fallback for old browsers */
   background: -webkit-linear-gradient(
     to right,
-    #38ef7d,
-    #11998e
+    #fbc7d4,
+    #9796f0
   ); /* Chrome 10-25, Safari 5.1-6 */
   background: linear-gradient(
     to right,
-    #38ef7d,
-    #11998e
+    #fbc7d4,
+    #9796f0
   ); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
 }
